fix: service 분리, print문 제거, test 코드 수정
